@@ -1,30 +1,15 @@
 {
   "main": "./app/app.ts",
   "dependencies": {
-    "@nativescript/android": "8.0.0",
-    "@nativescript/core": "8.0.0",
-    "@nativescript/ios": "8.0.0",
+    "@nativescript/core": "~8.1.0",
     "@nativescript/theme": "^3.0.1",
-    "nativescript-effects": "../src",
-    "nativescript-unit-test-runner": "0.7.0"
+    "nativescript-effects": "../src"
   },
   "devDependencies": {
-    "@nativescript/types": "8.0.0",
-    "@nativescript/webpack": "5.0.0-beta.2",
-    "babel-traverse": "6.12.0",
-    "babel-types": "6.11.1",
-    "babylon": "6.8.4",
-    "filewalker": "0.1.2",
-    "jasmine-core": "^2.5.2",
-    "karma": "^1.3.0",
-    "karma-jasmine": "^1.0.2",
-    "karma-nativescript-launcher": "^0.4.0",
-    "lazy": "1.0.11",
-    "nativescript-dev-typescript": "libs",
-    "tslint": "^6.1.3",
-    "typescript": "4.0.7"
-  },
-  "scripts": {
-    "ci.tslint": "npm i && tslint --config '../tslint.json' 'app/**/*.ts' --exclude '**/node_modules/**'"
+    "@nativescript/android": "~8.1.1",
+    "@nativescript/ios": "~8.1.0",
+    "@nativescript/types": "^8.1.0",
+    "@nativescript/webpack": "~5.0.0",
+    "typescript": "~4.5.4"
   }
 }
