@@ -49,21 +54,15 @@
     "license": "MIT",
     "homepage": "https://github.com/alexziskind1/nativescript-effects",
     "readmeFilename": "README.md",
-    "peerDependencies": {
-        "@nativescript/core": "^8.0.0"
-    },
     "devDependencies": {
-        "@nativescript/core": "8.0.0",
-        "@nativescript/types": "8.0.0",
-        "@nativescript/webpack": "5.0.0-beta.2",
-        "prompt": "^1.0.0",
-        "rimraf": "^2.5.0",
-        "ts-patch": "1.3.0",
+        "@nativescript/core": "^8.1.0",
+        "@nativescript/types": "^8.1.0",
+        "@nativescript/webpack": "~5.0.0",
+        "rimraf": "^3.0.2",
+        "ts-patch": "^1.3.0",
         "tslint": "^6.1.3",
-        "typescript": "4.0.7"
-    },
-    "dependencies": {
-        "ts-node": "^9.0.0"
+        "ts-node": "9.1.1",
+        "typescript": "~4.5.4"
     },
     "bootstrapper": "nativescript-plugin-seed"
 }
