@@ -1,5 +1,5 @@
 ﻿import { Application } from '@nativescript/core'
 
-const tnsfx = require('nativescript-effects');
+import 'nativescript-effects';
 
 Application.run({ moduleName: 'app-root' })
