@@ -1,2 +1,2 @@
 /// <reference path="./node_modules/@nativescript/types/index.d.ts" />
-/// <reference path="./node_modules/nativescript-effects/tns-effects.d.ts" />
+/// <reference path="../src/index.d.ts" />
